Refactor : move ModelPicker component outside of input bar files.
