+++ front/components/assistant/conversation/input_bar/InputBarModelPicker.tsx
@@ -1,53 +1,7 @@
 import { InputBarContext } from "@app/components/assistant/conversation/input_bar/InputBarContext";
-import { ModelPickerContent } from "@app/components/assistant/conversation/input_bar/ModelPickerContent";
-import type {
-  MakerGroup,
-  ModelTierId,
-  Selection,
-} from "@app/components/assistant/conversation/input_bar/modelPickerUtils";
-import {
-  buildModelSelection,
-  buildTierSelection,
-  getInitialEffort,
-  getModelEffortTier,
-  getModelTier,
-  getModelWithReasoningEffortLabel,
-  isPremiumModel,
-  isSameSelection,
-  isTierLocked,
-  resolveShownSelection,
-} from "@app/components/assistant/conversation/input_bar/modelPickerUtils";
-import { getModelMakerLogo } from "@app/components/providers/types";
-import { useTheme } from "@app/components/sparkle/ThemeContext";
-import { useAuth, useFeatureFlags } from "@app/lib/auth/AuthContext";
-import { useModels } from "@app/lib/swr/models";
-import type { AgentModelConfigurationType } from "@app/types/assistant/agent";
-import { isModelStreamId } from "@app/types/assistant/models/auto";
-import { getModelMaker } from "@app/types/assistant/models/providers";
-import type {
-  ModelConfigurationType,
-  ModelMakerIdType,
-  ModelSelectionType,
-  ReasoningEffort,
-} from "@app/types/assistant/models/types";
-import { isCreditPricedPlan } from "@app/types/plan";
-import type { LightWorkspaceType } from "@app/types/user";
-import {
-  BarFull,
-  BarHalf,
-  BarLow,
-  Button,
-  DropdownMenu,
-  DropdownMenuTrigger,
-} from "@dust-tt/sparkle";
-import type { ComponentType, MutableRefObject } from "react";
-import { useContext, useEffect, useMemo, useRef, useState } from "react";
-
-const TIER_BUTTON_ICON: Record<ModelTierId, ComponentType> = {
-  fast: BarLow,
-  standard: BarHalf,
-  complex: BarFull,
-};
+import type { ModelPickerProps } from "@app/components/model_picker/ModelPicker";
+import { ModelPicker } from "@app/components/model_picker/ModelPicker";
+import { useContext } from "react";
 
 type InputBarModelPickerProps = Omit<
   ModelPickerProps,
@@ -77,7 +31,7 @@ export function InputBarModelPicker({
       agentId={agentId}
       lastRequestedModel={lastRequestedModel}
       owner={owner}
-      buttonVariant={"ghost-secondary"}
+      buttonVariant="ghost-secondary"
       buttonSize={buttonSize}
       showLabel={false}
       side={side}
@@ -87,268 +41,5 @@ export function InputBarModelPicker({
       stickyModelOverride={stickyModelOverride}
       setStickyModelOverride={setStickyModelOverride}
     />
-  );
-}
-
-interface ModelPickerProps {
-  agentModel: AgentModelConfigurationType | null;
-  agentId: string | null;
-  lastRequestedModel: ModelSelectionType | null;
-  owner: LightWorkspaceType;
-  buttonVariant: "outline" | "ghost-secondary";
-  buttonSize: "xs" | "sm";
-  showLabel: boolean;
-  // Which side the dropdown opens toward. Mirrors the agent picker: "top" in an
-  // active conversation (input bar pinned to the bottom), "bottom" on the new
-  // conversation screen where there is room below.
-  side?: "top" | "bottom";
-  disabled?: boolean;
-  // Read-at-submit sink. The picker writes the current toSend here (including
-  // derived changes like agent switches); never triggers a parent re-render.
-  selectionRef?: MutableRefObject<ModelSelectionType | undefined>;
-  // Fired only on intentional user picks / revert — safe to setState.
-  onSelectionChange?: (modelSelection: ModelSelectionType | undefined) => void;
-  stickyModelOverride?: ModelSelectionType | undefined;
-  setStickyModelOverride?: (
-    modelSelection: ModelSelectionType | undefined
-  ) => void;
-}
-
-export function ModelPicker({
-  agentModel,
-  agentId,
-  lastRequestedModel,
-  owner,
-  buttonVariant,
-  buttonSize,
-  showLabel,
-  side = "top",
-  disabled,
-  selectionRef,
-  onSelectionChange,
-  stickyModelOverride,
-  setStickyModelOverride,
-}: ModelPickerProps) {
-  const { hasFeature } = useFeatureFlags();
-  const hasModelsPicker = hasFeature("models_picker");
-  const { subscription } = useAuth();
-  const lockPremiumEfforts = !isCreditPricedPlan(subscription.plan);
-
-  const { isDark } = useTheme();
-
-  const [isOpen, setIsOpen] = useState(false);
-  const [search, setSearch] = useState("");
-
-  // Inline-expansion state, only used on width-constrained clients.
-  const [moreModelsExpanded, setMoreModelsExpanded] = useState(false);
-  const [expandedMaker, setExpandedMaker] = useState<ModelMakerIdType | null>(
-    null
-  );
-
-  const [userOverride, setUserOverride] = useState<Selection | null>(null);
-
-  const { models } = useModels({
-    owner,
-    disabled: !hasModelsPicker,
-  });
-
-  const { shown: baseSelection, agentDefault } = useMemo(
-    () =>
-      resolveShownSelection({
-        agentModel,
-        lastRequestedModel,
-        sessionSticky: stickyModelOverride,
-        models,
-      }),
-    [agentModel, lastRequestedModel, stickyModelOverride, models]
-  );
-
-  // When the user switches which agent they address, discard any model override
-  // so the picker falls back to the newly-selected agent's own default.
-  const prevAgentIdRef = useRef(agentId);
-  useEffect(() => {
-    if (agentId === prevAgentIdRef.current) {
-      return;
-    }
-    prevAgentIdRef.current = agentId;
-    setUserOverride(null);
-    setStickyModelOverride?.(undefined);
-  }, [agentId, setStickyModelOverride]);
-
-  const shown: Selection = userOverride ?? baseSelection;
-  const shownModelSelection = shown.toSend;
-
-  // Keep the send-time ref current — including agent switches / sticky
-  // resolution where the user didn't pick anything. Mutating a ref during
-  // render is fine and avoids any parent re-render.
-  if (selectionRef) {
-    selectionRef.current = hasModelsPicker ? shownModelSelection : undefined;
-  }
-
-  const canRevert = !isSameSelection(shown.display, agentDefault.display);
-
-  // Concrete, selectable models (meta-models are surfaced as tiers instead).
-  const allModels = useMemo<ModelConfigurationType[]>(
-    () =>
-      models.filter(
-        (model) => !isModelStreamId(model.modelId) && model.isSelectable
-      ),
-    [models]
-  );
-
-  // Group models by maker, preserving first-seen order of both makers and
-  // models within each maker.
-  const makerGroups = useMemo<MakerGroup[]>(() => {
-    const groups = new Map<ModelMakerIdType, ModelConfigurationType[]>();
-    for (const model of allModels) {
-      const makerId = getModelMaker(model);
-      const existing = groups.get(makerId);
-      if (existing) {
-        existing.push(model);
-      } else {
-        groups.set(makerId, [model]);
-      }
-    }
-    return Array.from(groups.entries()).map(([makerId, makerModels]) => ({
-      makerId,
-      models: makerModels,
-    }));
-  }, [allModels]);
-
-  const commit = (selection: Selection) => {
-    if (isSameSelection(selection.display, agentDefault.display)) {
-      // Exactly the agent default: keep no override so we defer to the agent's
-      // own config (toSend undefined).
-      setStickyModelOverride?.(undefined);
-    } else {
-      setStickyModelOverride?.(selection.toSend);
-    }
-    setUserOverride(selection);
-    onSelectionChange?.(selection.toSend);
-  };
-
-  // Picking a concrete model (or nudging its effort slider) must keep the menu
-  // and its open submenus visible so the effort can still be adjusted. The
-  // click briefly moves focus/pointer in a way Radix treats as an
-  // interaction-outside and dismisses the (sub)menu; we record the pick time
-  // and veto the close that immediately follows it (see `onOpenChange` and the
-  // submenu guards in `ModelPickerMoreModels`).
-  const lastModelInteractionAtMsRef = useRef(0);
-
-  const shouldBlockDismiss = () =>
-    Date.now() - lastModelInteractionAtMsRef.current < 300;
-
-  const onSelectTier = (tierId: ModelTierId) => {
-    if (isTierLocked(tierId, { lockPremiumEfforts })) {
-      return;
-    }
-    commit({
-      display: { kind: "tier", tierId },
-      toSend: buildTierSelection(tierId),
-    });
-  };
-
-  const onSelectModel = (model: ModelConfigurationType) => {
-    if (isPremiumModel(model, { lockPremiumEfforts })) {
-      return;
-    }
-    lastModelInteractionAtMsRef.current = Date.now();
-    const effort = getInitialEffort(model, { lockPremiumEfforts });
-    commit({
-      display: { kind: "model", model, effort },
-      toSend: buildModelSelection(model, effort),
-    });
-  };
-
-  const onChangeEffort = (effort: ReasoningEffort) => {
-    if (shown.display.kind !== "model") {
-      return;
-    }
-    lastModelInteractionAtMsRef.current = Date.now();
-    const { model } = shown.display;
-    if (
-      lockPremiumEfforts &&
-      getModelEffortTier(model.modelId, effort) === "premium"
-    ) {
-      return;
-    }
-    commit({
-      display: { kind: "model", model, effort },
-      toSend: buildModelSelection(model, effort),
-    });
-  };
-
-  const onRevert = () => {
-    setUserOverride(agentDefault);
-    setStickyModelOverride?.(undefined);
-    onSelectionChange?.(undefined);
-  };
-
-  if (!hasModelsPicker) {
-    return null;
-  }
-
-  const buttonIcon =
-    shown.display.kind === "tier"
-      ? TIER_BUTTON_ICON[shown.display.tierId]
-      : getModelMakerLogo(getModelMaker(shown.display.model), isDark);
-
-  const label =
-    shown.display.kind === "tier"
-      ? getModelTier(shown.display.tierId).name
-      : getModelWithReasoningEffortLabel(shown.display);
-
-  return (
-    <DropdownMenu
-      open={isOpen}
-      onOpenChange={(open) => {
-        // Ignore the dismissal that a model/effort pick triggers, so the menu
-        // stays open. The window is short enough not to swallow a genuine
-        // click-outside a moment later.
-        if (!open && shouldBlockDismiss()) {
-          return;
-        }
-        setIsOpen(open);
-        if (open) {
-          setSearch("");
-          setMoreModelsExpanded(false);
-          setExpandedMaker(null);
-        }
-      }}
-    >
-      <DropdownMenuTrigger asChild>
-        <Button
-          className="px-2"
-          variant={buttonVariant}
-          size={buttonSize}
-          icon={buttonIcon}
-          label={showLabel ? label : undefined}
-          tooltip={showLabel ? undefined : label}
-          disabled={disabled}
-        />
-      </DropdownMenuTrigger>
-      <ModelPickerContent
-        side={side}
-        shouldBlockDismiss={shouldBlockDismiss}
-        shown={shown}
-        agentDefault={agentDefault}
-        canRevert={canRevert}
-        lockPremiumEfforts={lockPremiumEfforts}
-        makerGroups={makerGroups}
-        allModels={allModels}
-        search={search}
-        onSearchChange={setSearch}
-        moreModelsExpanded={moreModelsExpanded}
-        onToggleMoreModels={() => setMoreModelsExpanded((v) => !v)}
-        expandedMaker={expandedMaker}
-        onToggleMaker={(makerId) =>
-          setExpandedMaker((current) => (current === makerId ? null : makerId))
-        }
-        onSelectTier={onSelectTier}
-        onSelectModel={onSelectModel}
-        onChangeEffort={onChangeEffort}
-        onRevert={onRevert}
-      />
-    </DropdownMenu>
   );
 }
