+++ front/components/model_picker/ModelPickerSelectionIndicator.tsx
